+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on the [KeepAChangeLog] project.
 - [#526] Cleaned up extra claims from UserInfo with distributed claims
 - [#528] Fix faulty redirect_uri with query
 - [#532] Fix userinfo endpoint without auhtn_event in session
+- [#528] Fix faulty redirect_uri with query
+- [#498] Clean up replaced tokens on refresh and add Client.clean_tokens to clean old and replaced tokens
 
 ### Removed
 - [#494] Methods and functions deprecated in previous releases have been removed
@@ -40,6 +42,7 @@ The format is based on the [KeepAChangeLog] project.
 [#526]: https://github.com/OpenIDC/pyoidc/issues/526
 [#528]: https://github.com/OpenIDC/pyoidc/issues/528
 [#532]: https://github.com/OpenIDC/pyoidc/pull/532
+[#498]: https://github.com/OpenIDC/pyoidc/issues/498
 
 ## 0.13.1 [2018-04-06]